feature/SDL (#2)
* feature/sdl

Built win64 and linux amd64 version static of SDL2 library and copied to the lib directory of the project. Updated cmake to include correct version of the library based on build options.

* feature/sdl

Got basic SDL initialization working in Windows and Linux.

* feature/sdl

Window works in windows there is a mesa error in Linux using WSL2.